CRAS: alsa_io - Remove extra audio fetch.
During playback of two streams, if one is much higher latency than the
other, there was a slight optimization to fetch the audio early if
possible.  There was a bug in that this fetched audio from the low
latency stream the first time it ran, and since the fetches don't block
unless about to underrun, it didn't save anything.  Remove the extra
fetch.

BUG=chromium-os:35272
TEST=alsa_io_unittest, plus simultaneous playback of two streams one
with 10mS latency the other with 100mS latency, ensure no dropouts.

Change-Id: I2b5196babdcb68093d6383a0c80ce77c9ff74766
Signed-off-by: Dylan Reid <dgreid@chromium.org>