CRAS: Rework checks for input/output streams.
With the addition of loopback streams, the check for 'has_input' and
'has_output' need to be modified.  Now the checks are for if a stream
uses input or output hardware and if it is a loopback device.
Loopback streams won't use input hardware, but they will use input shm
regions.  This causes a need for a distinction between the checks.

BUG=chromium:247160
TEST=still builds, passes unittests and can play/record audio.

Change-Id: Ieb236920aead61785d3bb9b3af922a40370ff470
Signed-off-by: Dylan Reid <dgreid@chromium.org>