CRAS: audio_thread - select sleep time from input or output explicitly.
Make sure to set the sleep time based on the frame rate of the correct
device.  Either input or output must be open.  This change will allow
for the delay to be set for the loopback case that will be added.  The
addition of loopback will introduce a case where neither input or
output device is being used, but the thread will still run to feed
zeros to the loopback.

BUG=chromium:247160
TEST=audio_thread_unittest, playback and record audio on Link.

Change-Id: Ief8b9b8aefda25c991fec3bc59e1336cd2dca24f
Signed-off-by: Dylan Reid <dgreid@chromium.org>