CRAS: audio_thread - Increase sleep fuzz.
If there are two output threads that wake up close to each other, the
first to wake can cause the second to wait.  This changes the how close
to the wakeup time a stream has to be before it is processed.  Making
this longer will ensure that either the two streams are processed at the
same time, or the first finishes before the second needs to run.

BUG=none
TEST=play music and youtube videos.

Change-Id: I115eedf6ff05525a46907595f68bc043c0f200ff
Signed-off-by: Dylan Reid <dgreid@chromium.org>