CRAS: Dump audio thread info to test client.
Instead of leaving audio info in syslog, print it from the test client.
This will allow it to be included in feedback logs.  Doing this did
require that the nominal message size be increased, this won't affect
anything other than a slight increase in stack usage on the rare
occasion the server sends the client a message.

BUG=chromium:338965
TEST=cras_test_client --dump_a

Change-Id: I1b96af85845a0c3d78c7f46ff8728ce3dc4ca971
Signed-off-by: Dylan Reid <dgreid@chromium.org>