CRAS: a2dp_io - Reset frames count when buffer fills up.
The bt buffer count tries to track how many frames are written to the
device vs how many have been consumed.  This number will inevitably
drift over time because the clock on the device is not known.  To
avoid this problem, reset the count whenever the end of the send
buffer fills up.  When the buffer is full we know that exactly the
buffer length of audio has been written but not consumed.

BUG=chromium:
TEST=play youtube and listen for dropouts overnight.

Change-Id: Id3f29a78650a5bf534480fb7886033e953fe0903
Signed-off-by: Dylan Reid <dgreid@chromium.org>