CRAS: audio_thread - fix setting min_fill level before playback.
If one of the devices has read more samples than the others during a
previous wake, its offset into the shm region needs to be accounted
for when finding the minimum amount of frames a stream can provide.
Fixing this removes the need to update the count while writing because
an accurate minimum will have already been determined and is
guaranteed to be available for writing.

BUG=chromium:418833
TEST=Play audio to multiple devices at once on Blaze.

Change-Id: I8588fff7c60f9c7e190db57c1c1b17b33939a4f6
Signed-off-by: Dylan Reid <dgreid@chromium.org>