CRAS: add flag for hotword stream
Add two flags, one to allow more than a callback threshold of data to
be sent over shm.  The other is to allow a stream's timing to be
driven by an input device.  This will allow for a stream to be
notified when samples are ready even if the input isn't clocked
traditionally, such as hotword devices.

Change-Id: I5938d0e08eb480d80142422d91869e7c7b32c234
Signed-off-by: Dylan Reid <dgreid@chromium.org>