CRAS: alert - allow data arguments to alerts
Add the ability for an alert to be signaled with an optional data
parameter.  Alerts triggered in this way will behave differently.  If
they are signaled N times with data, they will be called N times.
Alerts without data will combine alerts if they happen rapidly
together, possibly only executing the callback once.

This will be used to allow data to be passed in to the callbacks so that
more callback systems can use the alert infrastructure instead of
open-coding it.

Change-Id: I96c5a0485dca8b16dbb0f45d3e5114ca1b491abb
Signed-off-by: Dylan Reid <dgreid@chromium.org>