wip: allow shm to be shared across threads
Split the header and audio buffer in to two different mmaps.
Make sure they don't overlap.
The lifetimes os the header and the buffer are independent of the
client because the mmap will live until it is unmapped. The shm FD can
be closed as soon as the needed memory is mapped.

Still TODO - convert the raw pointers to volatile memory regions so we
don't need to manually impl `Send` for Header and Buffer.

Change-Id: I79973d4a44b4f3db161d8c0afa18bff5d81f5b16