CRAS: metrics - reuse metics buffer
when logging each level, the same bufer was sent but with expanded
contents each time, instead log each from the beginning.

This fixes an issue where snprintf was using the return value
incorrectly. It returns the length of the format string that would be
printed, not the amount the snprintf limit caused to be printed.

Change-Id: I6c8af006f38ea4ce24f190eb4145674b9df34e0b